megacommit: the fifteenquel
Former-commit-id: 2de75761ebfac9de9bd41960e11e3e2c36a42c66

+++ WorldsEngine/Scripting/ScriptBindings/Vec3.hpp
@@ -11,13 +11,13 @@ namespace worlds {
             wrenEnsureSlots(vm, 1);
             glm::vec3* vPtr = (glm::vec3*)wrenGetSlotForeign(vm, 0);
 
-            wrenSetSlotDouble(vm, 0, (*vPtr)[wrenGetSlotDouble(vm, 1)]);
+            wrenSetSlotDouble(vm, 0, (*vPtr)[(int)wrenGetSlotDouble(vm, 1)]);
         }
 
         static void setComp(WrenVM* vm) {
             glm::vec3* vPtr = (glm::vec3*)wrenGetSlotForeign(vm, 0);
 
-            (*vPtr)[wrenGetSlotDouble(vm, 1)] = wrenGetSlotDouble(vm, 2);
+            (*vPtr)[(int)wrenGetSlotDouble(vm, 1)] = wrenGetSlotDouble(vm, 2);
         }
 
         static void setAll(WrenVM* vm) {
